chore: redesign the application page navbar
