+++ src/views/pages/ApplicationDetails/ResourceStats.vue
@@ -1,0 +1,5 @@
+<script setup></script>
+
+<template>resource stats</template>
+
+<style scoped></style>
